% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_access}
\alias{cache_access}
\title{Gets information on the specified (cache) file access}
\usage{
cache_access(cache_path)
}
\arguments{
\item{cache_path}{A character scalar.}
}
\value{
A list.
}
\description{
Utility function \code{cache_access} returns a list ...
}
\examples{
\dontrun{cache_access("cache.rds")}
}

