% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_diseaseform_values.R
\name{get_diseaseform_values}
\alias{get_diseaseform_values}
\title{Fetches the values used to submit the search form}
\usage{
get_diseaseform_values(disease = character(), print_only = FALSE,
  set_global_only = FALSE, new_download = FALSE,
  file = "diseaseform_values.rds")
}
\arguments{
\item{disease}{A character scalar. The disease entered is matched against the
form labels by the \code{grep} function. It is case insensitive and can be
entered partially.}

\item{print_only}{A logical scalar. When \code{TRUE}, the result is printed
to the console and \code{NULL} returned invisibly.}

\item{set_global_only}{A logical scalar. When \code{TRUE}, \code{NULL} is
returned invisibly.}

\item{new_download}{A logical scalar. If \code{TRUE}, the cached data (either
supplied with the package or in the \code{file} argument) are ignored and
the data are fetched from the web. It is set to \code{TRUE} inside the
function if \code{years_available} in the cached data do not extend to
current year.}

\item{file}{A character scalar. If the file specified exists, it is used
  instead of the dataset supplied with the package. If \code{new_download} is
  \code{TRUE}, the cache is written to the file specified.}
}
\value{
An object of class \code{c("tbl", "tbl_df", "data.frame", "diseaseform")}
  containing values for all diseases available if \code{disease} was not
  specified, or containing exactly one row when \code{disease} was specified
  and matched.

\code{NULL} (invisibly) when \code{set_global_only} is set to
  \code{TRUE} or (invisibly) when \code{print_only} is set to \code{TRUE} or
  (with a warning) when \code{disease} does not give exactly one match.
}
\description{
Utility function \code{get_diseaseform_values} fetches the values needed for
submitting the initial search form at
\href{http://www.oie.int/wahis_2/public/wahid.php/Diseaseinformation/Immsummary}{WAHID
--> Disease information --> Immediate notifications and Follow-ups} -- the
\emph{disease_id_hidden} and \emph{disease_type_hidden} for a name of the
disease entered. It also fetches the range of years available.
}
\details{
The function utilizes a cache that is as a default supplied as a dataset
\code{data(diseaseform_values)}, but optionally a path to a cache file can be
supplied.
}
\section{Side effects}{
 This function sets the variables \code{values_labels}
  containing the form value for all diseases available and
  \code{years_available} in the \code{globals} environment that are utilized
  by other functions from the package. As mentioned above, it also optionally
  writes a cache file at the location specified by the user.
}
\examples{
\dontrun{get_diseaseform_values(disease = "african swine fever")}
}
\seealso{
\code{\link{read_cache}}, \code{\link{write_cache}}
}

