% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_sinf.R
\name{get_sinf}
\alias{get_sinf}
\alias{get_sinf.character}
\alias{get_sinf.diseaseform}
\alias{get_sinf.numeric}
\title{Downloads and processes the Summaries of Immediate notifications and
Follow-ups.}
\usage{
get_sinf(x, years, countries, coutries_match = "all", new_download = FALSE,
  file = "sinf_cache.rds", cache_interval = "1 day",
  suppress_messages = FALSE, ...)

\method{get_sinf}{numeric}(x, years, countries, countries_match = "all",
  new_download = FALSE, file = "sinf_cache.rds", cache_interval = "1 day",
  suppress_messages = FALSE)

\method{get_sinf}{character}(x, years, countries, countries_match = "all",
  new_download = FALSE, file = "sinf_cache.rds", cache_interval = "1 day",
  suppress_messages = FALSE)

\method{get_sinf}{diseaseform}(x, years, countries, countries_match = "all",
  new_download = FALSE, file = "sinf_cache.rds", cache_interval = "1 day",
  check_x = TRUE, suppress_messages = FALSE)
}
\arguments{
\item{x}{Specifies the disease of interest. See methods.}

\item{years}{An integer vector specifying the years of interest. Defaults to
the full range available (2005 - ) if not specified.}

\item{countries}{A character vector specifying the countries or areas of
interest (see \code{\link{match_countries}}). Defaults to whole world if
not specified.}

\item{new_download}{A logical scalar. If \code{TRUE}, the cached data (either
supplied with the package or in the \code{file} argument) are ignored and
the data are fetched from the web.}

\item{file}{A character scalar. Path to the cache file (see
\code{\link{read_cache}}, \code{\link{write_cache}}).}

\item{cache_interval}{A specification of a duration (default \code{"1 day"})
or a date (see \code{\link{cache_cut_off}}). Older records of uncomplete
years (see details) are downloaded, records of uncomplete years within the
cache interval and record of complete years are taken from the \code{file}
or the package-supplied data.}

\item{suppress_messages}{A logical scalar. If \code{TRUE}, all informative
  messages from the function are suppressed.}

\item{countries_match}{A character vector containing any combination of
column names to match \code{countries} against in
\code{countries_available} (see \code{\link{get_countries}}), or
\code{"all"} (the default). Ignored if \code{countries} not specified.}

\item{check_x}{A logical scalar specifying whether the disease given as
  \code{diseaseform} should be checked. Set to \code{FALSE} when
  \code{get_sinf.diseaseform} called from other methods.}
}
\value{
An object of class \code{c("sinf", "tbl", "tbl_df", "data.frame")}
  containing summaries of immediate notifications and follow-ups from the OIE
  WAHID website for the disease, years, and countries given.

\code{NULL} if anything specified incorrectly or not matched.
}
\description{
High-level generic function \code{get_sinf} fetches and processes the
Summaries of immediate notifications and Follow-ups for specified disease,
years, and countries from the OIE WAHID website
(\href{http://www.oie.int/wahis_2/public/wahid.php/Diseaseinformation/Immsummary}{WAHID
--> Disease information --> Immediate notifications and Follow-ups}).
}
\details{
The function utilises data supplied with the package (\code{data(SINF)}) and
also a local cache to minimise OIE website traffic. Information on the
website is organised by disease and year. Records for a given year older than
\code{cache_interval} are considered expired if any of the records has status
"Continuing" or if that year is the current year or the last year (if run in
the first quarter).
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: 

\item \code{character}: 

\item \code{diseaseform}: 
}}
\section{Side effects}{
 This function sets variables \code{web_not_changed},
  \code{values_labels}, \code{years_available}, and
  \code{countries_available} (if not set previously) in the \code{globals}
  environment that are utilized by other functions from the package. As
  mentioned above, it also writes a cache file at the location specified by
  the user.
}
\examples{
\dontrun{get_sinf("african swine fever")}
}

