% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countries.R
\name{match_countries}
\alias{match_countries}
\alias{match_countries.character}
\title{Matches patterns entered against the table of countries available on the OIE
WAHID website}
\usage{
match_countries(x, fields = "all", select = "ISO3", simplify = TRUE)

\method{match_countries}{character}(x, fields = "all", select = "ISO3",
  simplify = TRUE)
}
\arguments{
\item{x}{A \code{character} scalar or vector containing a pattern or patterns
to search for.}

\item{fields}{A \code{character} scalar or vector giving the names of fields
to match against. Possible values are "all" (default) or any combination of
column names of the \code{countries_available}.}

\item{select}{A \code{character} scalar or vector giving the names of fields
to be returned. Possible values are "all" or any combination of column
names of the \code{countries_available} (default "ISO3").}

\item{simplify}{A \code{logical} scalar specifying whether the result should
  be simplified to character vector if only one field has to be returned.}
}
\value{
A \code{character} vector or \code{tbl_df} containing the selected
  columns from the \code{countries_available} for records that were matched
  by the pattern(s) entered. \code{NULL} if there is no match or any of the
  arguments is specified incorrectly.
}
\description{
Utility function \code{match_countries} is used by other functions from the
package to match country or area specification against the
\code{countries_available} table.
}
\details{
The country or area specification can be entered as a pattern or a character
vector of patterns. In default, it is matched against all fields in the
\code{countries_available} table and ISO3 codes of countries matched are
returned as \code{character} vector. Matching is case-insensitive. Fields to
match against and fields returned can be specified. Multiple fields are
returned as \code{tbl_df}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}
\examples{
match_countries("Europe", fields = "OIE_region")
match_countries("^East.*Europe", select = "all")
match_countries("Caribbean", fields = "GEO3")
match_countries("Congo", select = "all")
match_countries("Kongo")
}

