% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{read_cache}
\alias{read_cache}
\title{Deserializes an R object from the specified (cache) file (in RDS format)}
\usage{
read_cache(cache_path)
}
\arguments{
\item{cache_path}{A character scalar.}
}
\value{
An R object deserialized from the RDS file or \code{NULL}.
}
\description{
Utility function \code{read_cache} returns the R object serialized in the
file specified. If the file does not exist, is not readable, or is not
a RDS file, \code{NULL} is returned and a message printed to console.
}
\examples{
\dontrun{read_cache("cache.rds")}
}

