% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wahid2r.R
\docType{package}
\name{wahid2r}
\alias{wahid2r}
\alias{wahid2r-package}
\title{wahid2r: A package for scraping data on the occurence of OIE listed animal
diseases from the OIE WAHID website.}
\description{
The wahid2r package provides functions to download Summaries of Immediate
notifications and Follow-ups, Summaries of outbreaks, and Outbreak reports on
all OIE-listed animal diseases (since 2005) from the OIE WAHID website
(\url{http://www.oie.int/wahis_2/public/wahid.php/Diseaseinformation/Immsummary}).
}
\section{Warning}{
 The data might be incomplete, depending on the quality of
  reporting by the individual countries. Use only for information purposes.
}

