% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{write_cache}
\alias{write_cache}
\title{Serializes the specified object into the specified (cache) file (RDS format)}
\usage{
write_cache(x, cache_path)
}
\arguments{
\item{x}{The R object to be written.}

\item{cache_path}{The file path.}
}
\value{
The file path written (a character scalar) or \code{NULL}.
}
\description{
Utility function \code{write_cache} writes the specified object in serialized
(RDS) format to the specified file. If the file cannot be written or created,
tries to write to a file with the same name in the working directory. If
successful, returns the path to the file written, otherwise returns
\code{NULL}.
}
\examples{
\dontrun{write_cache(x, "cache.rds")}
}

