% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_web.R
\name{check_web}
\alias{check_web}
\title{Checks if the OIE WAHID website has not changed.}
\usage{
check_web(hash_file = "hashes.rds", write_cache_if_different = FALSE)
}
\arguments{
\item{hash_file}{A character scalar.}

\item{write_chache_if_different}{A logical scalar.}
}
\value{
\code{TRUE} If critical features of the web have not changed compared
  to the cached state.

\code{FALSE} If critical features of the the web have changed.
}
\description{
Utility function \code{check_web} downloads the
\href{http://www.oie.int/wahis_2/public/wahid.php/Diseaseinformation/Immsummary}{WAHID
--> Disease information --> Immediate notifications and Follow-ups} webpage
and compares computed hashes of some critical features with those stored in
cache. Returns \code{TRUE} if the hashes are identical, otherwise returns
\code{FALSE} with a diagnostic message indicating the features that have
changed. The cache is as a default supplied as a dataset \code{data(hashes)},
but optionally a path to a cache file can be supplied, and also an option to
write such file when the hashes are not identical.
}
\examples{
\dontrun{check_web()}
}

